pyproject.toml
[tool.poetry]
name = "psra-origin-engine"
version = "1.0.0"
description = "LangGraph-powered Origin Calculation Engine for PSRA-LTSD"
authors = ["Sevensa <info@sevensa.nl>"]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.1.11"
langchain = "^0.1.0"
langchain-openai = "^0.1.0"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
tenacity = "^8.2.3"
asyncpg = "^0.29.0"
aiohttp = "^3.9.0"
redis = "^5.0.0"
psutil = "^5.9.0"
great-expectations = "^0.18.12"
pandas = "^2.2.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.1.0"
maturin = "^1.4.0"
pre-commit = "^3.6.0"
testcontainers = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint]
extend-select = ["Q", "RUF"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
